requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - Supabase Client
supabase==2.3.4
postgrest==0.13.2

# Database - SQLAlchemy (kept temporarily for non-migrated endpoints)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
pyjwt==2.8.0
cryptography==42.0.0
passlib[bcrypt]==1.7.4

# Task Queue
celery==5.3.6
redis==5.0.1

# Data Processing
pandas==2.2.0
pyarrow==15.0.0
openpyxl==3.1.2
xlrd==2.0.1
umap-learn==0.5.5

# HTTP Client
httpx>=0.24,<0.26

# AI/LLM - Local inference
# Used for local AI interpretation with Ollama (no data export)
# Install Ollama separately: brew install ollama (macOS) or see https://ollama.ai

# File Processing
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.4.0
email-validator>=2.0.0
scipy>=1.12.0
fastcluster>=1.2.6

scipy>=1.12.0
fastcluster>=1.2.6
goatools>=1.3.1
